+++ CHANGELOG.md
@@ -11,8 +11,9 @@
 - [TD-3453] Purge logically deleted structures
 - [TD-3906] Notes audit now includes status changes
 - [TD-3050] Show quality errors
+- [TD-3945] Created Grant entity
+- [TD-3947] Display user grant in data structure
 - [TD-3551] Restore rule implementations
-- [TD-3945] Created `Grant` entity
 
 ## [4.24.0] 2021-07-13